# src/main/resources/api_specs/api-specs.yaml
openapi: 3.0.0
info:
  title: Product API
  description: API for managing products
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
paths:
  /products:
    get:
      summary: Retrieve all products
      operationId: getAll
      tags:
        - Products
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
  
    post:
      summary: Create a new product
      operationId: create
      tags:
        - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
      responses:
        '201':
          description: Created product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
  
  /products/{productId}:
    get:
      summary: Retrieve a product by ID
      operationId: getById
      tags:
        - Products
      parameters:
        - name: productId
          in: path
          required: true
          description: ID of the product to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Product not found
  
    put:
      summary: Update an existing product
      operationId: update
      tags:
        - Products
      parameters:
        - name: productId
          in: path
          required: true
          description: ID of the product to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
      responses:
        '200':
          description: Updated product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Product not found
  
    delete:
      summary: Delete a product
      operationId: delete
      tags:
        - Products
      parameters:
        - name: productId
          in: path
          required: true
          description: ID of the product to delete
          schema:
            type: integer
      responses:
        '200':
          description: Product successfully deleted

components:
  schemas:
    ProductDTO:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          description: The ID of the product
        title:
          type: string
          description: The title of the product
        description:
          type: string
          description: The description of the product
        categoryId:
          type: integer
          format: int64
          description: The category ID of the product
        price:
          type: number
          format: double
          description: The price of the product
        currency:
          type: string
          description: The currency of the product price
        stock:
          type: integer
          description: The available stock count of the product
        manufacturer:
          type: string
          description: The manufacturer of the product